mudanca de tipo em atributos da classe PokeDetalhe.kt para double

+++ app/src/main/java/com/eriberto/pokedex/repository/model/PokeDetalhe.kt
@@ -3,8 +3,8 @@ package com.eriberto.pokedex.repository.model
 data class PokeDetalhe(
     val name: String?,
     val types: List<TypeSlot>,
-    val weight: Int,
-    val height: Int,
+    val weight: Double,
+    val height: Double,
     val abilities: List<AbilitySlot>,
     val game_indices: List<Game>
 )
